+++ src/evm-lite-core/src/index.ts
@@ -2,10 +2,10 @@
 export { default as EVMLC } from './classes/EVMLC';
 export { default as Account } from './classes/accounts/Account';
 export { default as Accounts } from './classes/accounts/Accounts';
-export {
-	default as SolidityContract
-} from './classes/contract/SolidityContract';
+export { default as Contract } from './classes/contract/Contract';
 export { default as Transaction } from './classes/transaction/Transaction';
+
+export { ABI } from './classes/contract/Contract';
 
 // Types
 export { TXReceipt } from './clients/TransactionClient';
@@ -14,10 +14,7 @@ export {
 	SentTransaction as SentTX,
 	SignedTransaction
 } from './classes/transaction/Transaction';
-export {
-	BaseContractSchema,
-	ContractABI
-} from './classes/contract/SolidityContract';
+export { BaseContractSchema, ContractABI } from './classes/contract/Contract';
 export { V3JSONKeyStore } from './classes/accounts/Account';
 
 export * from './types/index';