+++ CHANGELOG.md
@@ -13,6 +13,8 @@ IMPROVEMENTS:
 - Renamed `Keystore.decryptAccount` to `Keystore.decrypt`.
 - Removed one param from `Keystore.list`. Now only accepts one param
   representing a `EVMLC` object to fetch balance and nonce from the node.
+- `Transaction` submittion now only signed one way requiring non optional
+  `account` object
 
 SECURITY: